split up files in corehq/apps/fixtures/upload/
