+++ corehq/apps/fixtures/upload/__init__.py
@@ -1,5 +1,6 @@
+from .const import DELETE_HEADER
+from .validation import validate_fixture_file_format
 from .run_upload import upload_fixture_file
-from .upload import DELETE_HEADER, validate_fixture_file_format
 
 __all__ = [
     'DELETE_HEADER',
