+++ corehq/apps/fixtures/upload/const.py
@@ -1,0 +1,1 @@
+DELETE_HEADER = "Delete(Y/N)"
