+++ corehq/apps/fixtures/upload/definitions.py
@@ -1,0 +1,21 @@
+import re
+from django.template.loader import render_to_string
+
+
+class FixtureUploadResult(object):
+    """
+    Helper structure for handling the results of a fixture upload.
+    """
+
+    def __init__(self):
+        self.success = True
+        self.messages = []
+        self.errors = []
+        self.number_of_fixtures = 0
+
+    def get_display_message(self):
+        message = render_to_string('fixtures/partials/fixture_upload_status_api.txt', {
+            'result': self,
+        })
+        message = u'\n'.join(re.split(r'\n*', message)).strip()
+        return message
